Gem housekeeping, pt. 1
- Update gemspec information

- Organize dependencies into .gemspec and .gemfile-s

- Add scripts/nats-server to automatically install and run nats-server from a temporary folder

- Update RSpec configuration (organize helpers, random order)

- Fix some flaky specs
