+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,24 @@
+<!--
+  First of all, thanks for contributing!
+
+  If it's a typo fix or minor documentation update feel free to skip the rest of this template!
+-->
+
+## What is the purpose of this pull request?
+
+<!--
+  If it's a bug fix, then link it to the issue, for example:
+
+  Fixes #xxx
+-->
+
+## What changes did you make? (overview)
+
+## Is there anything you'd like reviewers to focus on?
+
+## Checklist
+
+- [ ] I've added tests for this change
+- [ ] I've added a Changelog entry
+- [ ] I've updated a documentation
+
