+++ .gitignore
@@ -1,3 +1,45 @@
+# Numerous always-ignore extensions
+*.diff
+*.err
+*.orig
+*.log
+*.rej
+*.swo
+*.swp
+*.vi
+*~
+*.sass-cache
 *.iml
+.idea/
+
+# Sublime
+*.sublime-project
+*.sublime-workspace
+
+# OS or Editor folders
 .DS_Store
-/.idea
+.cache
+.project
+.settings
+.tmproj
+Thumbs.db
+
+.bundle/
+log/*.log
+pkg/
+spec/dummy/db/*.sqlite3
+spec/dummy/db/*.sqlite3-journal
+spec/dummy/tmp/
+
+Gemfile.lock
+Gemfile.local*
+.rspec-local
+.ruby-version
+*.gem
+
+tmp/
+.rbnext/
+
+gemfiles/*.lock
+
+dist/
