--- TODO.md
@@ -1,24 +1,0 @@
-
-
-- [X] pub
-- [X] sub
-- [X] unsub
-- [X] ping/pong interval
-- [X] spec
-- [X] request/response
-- [X] authorization
-- [X] reconnection
-- [X] benchmark scripts
-- [X] process op errors from read/parser
-- [X] reading deadlines
-- [X] graceful close
-- [X] error/disconnect/close callbacks
-- [X] cluster support
-- [X] TLS
-- [X] travis
-- [X] auto discovery
-- [X] gem release
-- [ ] increase test coverage
-- [ ] max payload err handling
-- [ ] read dynamic socket buffer
-- [X] nuid
