+++ gemfiles/rails6.gemfile
@@ -1,0 +1,7 @@
+source "https://rubygems.org"
+
+gem "rails", "~> 6.0"
+gem "sqlite3", "~> 1.4"
+
+gemspec path: ".."
+
