+++ gemfiles/rails8.gemfile
@@ -1,0 +1,6 @@
+source "https://rubygems.org"
+
+gem "rails", "~> 7.2"
+gem "sqlite3", "~> 2.0"
+
+gemspec path: ".."
