+++ gemfiles/railsmain.gemfile
@@ -1,0 +1,6 @@
+source "https://rubygems.org"
+
+gem "rails", github: "rails/rails"
+gem "sqlite3"
+
+gemspec path: ".."
