+++ lib/nats-pure.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require "nats/io/client"
