--- spec/.rspec
@@ -1,2 +1,0 @@
--fd
--c
