+++ spec/client_tls_spec.rb
@@ -16,6 +16,20 @@ require 'spec_helper'
 require 'openssl'
 require 'erb'
 
+DEFAULT_JRUBY_CIPHER_SUITE = %q(
+    # JRuby is sensible to the ciphers being used
+    # so we specify the ones that are available on it here.
+    # See: https://github.com/jruby/jruby/issues/1738
+    cipher_suites: [
+      "TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256",
+      "TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA",
+      "TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA",
+      "TLS_RSA_WITH_AES_128_CBC_SHA",
+      "TLS_RSA_WITH_AES_256_CBC_SHA",
+      "TLS_RSA_WITH_3DES_EDE_CBC_SHA"
+    ]
+)
+
 describe 'Client - TLS spec' do
 
   context 'when server requires TLS and no auth needed' do
@@ -134,7 +148,7 @@ describe 'Client - TLS spec' do
       nats.connect('tls://127.0.0.1:4444', reconnect:false) rescue nil
       expect(nats.options[:tls]).to_not be_nil
     end
-    
+
     it 'should allow custom secure connection contexts' do
       errors = []
       closes = 0
